hf_space/requirements.txt
gradio>=4.44.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
numpy>=1.23.0
